requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
dnspython==2.4.2
flask==3.0.0
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
m-kafka-sdk-v2==0.1.8
MarkupSafe==2.1.3
pymongo==4.6.0
requests==2.31.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
